requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
defusedxml==0.7.0rc1
dj-database-url==0.5.0
dj-rest-auth==1.1.1
Django==3.0.8
django-environ==0.4.5
django-heroku==0.3.1
djangorestframework==3.11.0
django-cors-headers
gunicorn==20.0.4
psycopg2==2.8.5
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.2.0
django_filter==2.4.0
coreapi
django-rest-swagger